+++ src/stratega/src/ForwardModel/SamplingMethod.cpp
@@ -177,7 +177,7 @@ std::vector<SGA::Vector2i> SGA::Dijkstra::getPositions(const GameState& gameStat
 			auto isValidPos = [&](int x, int y, float totalCost)
 			{
 				if (gameState.getTileAt({ x, y }).getTileTypeID() == -1 || !gameState.getTileAt({ x, y }).isWalkable()
-					|| std::floor(totalCost + 1) > searchSize)
+					|| std::floor(totalCost + 1) > (float)searchSize)
 					return false;		
 				else
 					return true;
